+++ apps/ui/src/features/docs/components/data-table-demo.tsx
@@ -1,5 +1,6 @@
 "use client"
 
+/* eslint-disable react-refresh/only-export-components, react-hooks/incompatible-library */
 import * as React from "react"
 import {
   flexRender,